add a parent route
