+++ app/styles/_my_dashboard.scss
@@ -18,6 +18,7 @@
 
   .tabs .tab {
     text-align: center;
+    color: white;
     opacity: 0.5;
     cursor: pointer;
